requirements.txt
Flask==2.3.2
flask_cors
requests==2.31.0
numpy
Pillow==9.5.0
torch
torchvision
facenet-pytorch
gunicorn
